% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pop_dyn}
\alias{example_pop_dyn}
\title{Example of population dynamic parameters as provided by users}
\format{
A data frame with 10 rows and 5 variables: \describe{
\item{locality}{species locality}
\item{species}{species identifying string}
\item{lifestage}{lifestage}
\item{reproduction}{reproduction rate}
\item{survival}{survival rate}}
}
\usage{
example_pop_dyn
}
\description{
A dataset containing reproduction and survival rates, for different species
at different lifestages
}
\keyword{datasets}
