% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_evo.R
\name{fetch_survival}
\alias{fetch_survival}
\title{Helper Function Fetch a survival value for a specific lifestage from a
dynamic_species intermediary dataframe}
\usage{
fetch_survival(input_df, selected_lifestage)
}
\arguments{
\item{input_df}{Input dataframe that contains at least the columns survival
and lifestage, ideally one that's been filtered already to a specific
species and location}

\item{selected_lifestage}{The lifestage to get, actually \code{dplyr::pull}
, the survival value from}
}
\value{
one or more values of the same class as in the input dataframe
}
\description{
Helper Function Fetch a survival value for a specific lifestage from a
dynamic_species intermediary dataframe
}
\examples{

data.table::fread(file.path("data", "pop_dyn.csv")) \%>\%
dplyr::filter(species == "deer",
locality == "Wallonia") \%>\%
fetch_locality
}
