% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_evo.R
\name{viz_pop_evo}
\alias{viz_pop_evo}
\title{Build a population matrix, project the evolution of the population, and
visualise the results}
\usage{
viz_pop_evo(
  input_df,
  selected_species,
  selected_locality,
  n = NULL,
  years = 10,
  colours = NULL,
  show_labels = TRUE
)
}
\arguments{
\item{input_df}{Input dataframe, with columns \code{locality}, \code{species}
, \code{lifestage}, \code{reproduction} and \code{survival}}

\item{selected_species}{The species from the input_df for which the model
should be visualized}

\item{selected_locality}{The locality for the species that should be
visualized}

\item{n}{A vector of integers, specifying the number of individuals per
lifestage. If not provided, the function will default to 100 individuals for
every lifestage}

\item{years}{The number of years the model should run for, really the number
of equally spaced points in time between the lifestages}

\item{colours}{A character vector of hex codes of equal length to the number
of lifestages, these colours are used in the plot this function generates.}

\item{show_labels}{Either \code{TRUE} or \code{FALSE}, optionally show extra
labels on the output plot. This might be useful if no colours are provided}
}
\value{
a ggplot object that shows the projection of the population for
\code{years} steps
}
\description{
Build a population matrix, project the evolution of the population, and
visualise the results
}
\examples{
```
data <- data.table::fread(file.path("data", "pop_dyn.csv"))
viz_pop_evo(data,
            "wild boar",
            "Flanders",
            colours = colstring_to_hex("cyan","yellow2","plum4"),
             show_labels = FALSE)
```
}
\seealso{
\code{\link[=colstring_to_hex]{colstring_to_hex()}} which can be used to convert built-in R colours
to hex codes. If no colours are provided (the default), the function will
return all resulting curves in black
}
