% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizpopevo-package.R
\docType{package}
\name{vizpopevo-package}
\alias{vizpopevo}
\alias{vizpopevo-package}
\title{vizpopevo: vizpopevo: From a standardized input, filter and create a population matrix, project the population evolution and plot the model output}
\description{
This package provides one main function: `viz_pop_evo()`, and a number of helpers. Currently the main function really does a bit much, it loads and filters the input data, transforms it to a population matrix, runs the model and returns a visualization of this model. It offers a number of defaults for optional arguments, and will warn you when required arguments are missing. It also checks the input dataframe. It offers a small amount of customizability in regards to the output plot, you can set the colours of the resulting curves, and you can turn curve labels on or off.
}
\author{
\strong{Maintainer}: Pieter Huybrechts \email{pieter.huybrechts@meisebotanicgarden.com} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID})

}
\keyword{internal}
